Refactor ``Expr`` class
