+++ dask_expr/tests/test_datasets.py
@@ -2,7 +2,7 @@ from dask.dataframe.utils import assert_eq
 
 from dask_expr import new_collection
 from dask_expr.datasets import timeseries
-from dask_expr.expr import Lengths
+from dask_expr.frameexpr import Lengths
 
 
 def test_timeseries():
